# ARROW-15995: [GO] Improve 'sum_float64_neon' performance