Improve the 'Float64Funcs_Sum_1000000-64':

From:
BenchmarkFloat64Funcs_Sum_1000000-64                1179           1016745 ns/op        7868.25 MB/s
to:
BenchmarkFloat64Funcs_Sum_1000000-64                1387            849128 ns/op        9421.43 MB/s

Get 19.7% performance uplift in cases which data set is larger than 1MB.


Test:
```
ARM_ENABLE_EXT=NEON go test
PASS
ok      github.com/apache/arrow/go/v8/arrow/math        0.010s
```